refactor(engine): rename l2-team plugin to agent-team and split modules
