+++ packages/engine/src/built-in/agent-team-plugin/constants.ts
@@ -1,0 +1,6 @@
+export const DEFAULT_MAX_PARALLEL = 3;
+export const MAX_PARALLEL_LIMIT = 20;
+export const DEFAULT_NODE_TIMEOUT_MS = 120_000;
+
+export const TEAM_RUN_TOOL_NAME = 'team_run';
+export const AGENT_TEAM_REGISTRY_SERVICE_KEY = 'agent-team:registry';
